[IMP] Direct informational messages to stderr.
This allows the GEDCOM output to be piped to another program without
polluting it with progress messages.

The usage text is now also directed to stderr, as is normal practice.

TESTED:
Piped output to an editor, with the progess messages still appearing
on the terminal.
Usage text now appears on the terminal, even with redirected stdout.